refactor: 불필요 ã주석 제거 및 Format static 함수 추가
