+++ src/main/java/com/ss/hanarowa/domain/branch/controller/BranchController.java
@@ -3,10 +3,7 @@ package com.ss.hanarowa.domain.branch.controller;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,7 +13,6 @@ import com.ss.hanarowa.global.response.ApiResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -34,7 +30,7 @@ public class BranchController {
 	 */
 	@GetMapping("")
 	@Operation(summary = "지점 목록 조회 API", description = "하나로와에서 이용할 수 있는 지점을 조회합니다.")
-	public ResponseEntity<ApiResponse<List<BranchResponseDTO>>> getBranchs() {
+	public ResponseEntity<ApiResponse<List<BranchResponseDTO>>> getBranches() {
 		List<BranchResponseDTO> branchList = branchService.getAllBranches();
 		return ResponseEntity.ok(ApiResponse.onSuccess(branchList));
 	}
