+++ src/main/java/com/ss/hanarowa/domain/facility/dto/reponse/FacilityDetailResponseDTO.java
@@ -2,10 +2,6 @@ package com.ss.hanarowa.domain.facility.dto.reponse;
 
 import java.util.List;
 import java.util.Map;
-
-import com.ss.hanarowa.domain.facility.entity.FacilityImage;
-import com.ss.hanarowa.domain.facility.entity.FacilityTime;
-
 import lombok.Builder;
 import lombok.Getter;
