+++ src/main/java/com/ss/hanarowa/domain/facility/dto/request/FacilityReservationDTO.java
@@ -1,9 +1,5 @@
 package com.ss.hanarowa.domain.facility.dto.request;
 
-import java.time.LocalDateTime;
-
-import com.ss.hanarowa.global.entity.BaseEntity;
-
 import lombok.Builder;
 import lombok.Getter;
