+++ src/main/java/com/ss/hanarowa/domain/facility/service/FacilityService.java
@@ -1,10 +1,6 @@
 package com.ss.hanarowa.domain.facility.service;
 
 import java.util.List;
-
-import org.springframework.security.core.Authentication;
-
-import com.ss.hanarowa.domain.facility.dto.reponse.AdminFacilityResponseDTO;
 import com.ss.hanarowa.domain.facility.dto.reponse.FacilityDetailResponseDTO;
 import com.ss.hanarowa.domain.facility.dto.reponse.FacilityReservationResponseDTO;
 import com.ss.hanarowa.domain.facility.dto.reponse.FacilityResponseDTO;
@@ -18,8 +14,6 @@ public interface FacilityService {
 	void deleteFacilityReservation(Long reservationId);
 
 	void reservateFacility(FacilityReservationDTO facilityReservationDTO, Long memberId);
-	
-	List<AdminFacilityResponseDTO> getAllFacilityReservations();
 
 	List<FacilityReservationResponseDTO> getAllMyFacilityReservations(String email);
 }
