+++ src/main/java/com/ss/hanarowa/domain/lesson/dto/response/LessonListSearchResponseDTO.java
@@ -1,7 +1,5 @@
 package com.ss.hanarowa.domain.lesson.dto.response;
 
-import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
