+++ src/main/java/com/ss/hanarowa/domain/lesson/repository/ReviewRepository.java
@@ -1,7 +1,6 @@
 package com.ss.hanarowa.domain.lesson.repository;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
