+++ src/main/java/com/ss/hanarowa/domain/member/dto/request/LoginRequestDTO.java
@@ -1,6 +1,5 @@
 package com.ss.hanarowa.domain.member.dto.request;
 
-
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
@@ -8,7 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 
 @Getter
 @Setter
