+++ src/main/java/com/ss/hanarowa/domain/member/dto/response/MemberListResponseDTO.java
@@ -9,7 +9,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
 @Getter
 @Setter
 @AllArgsConstructor
